create seperate store for the recently used keys
