+++ web/src/features/search/components/FilterBuilder/TagSelector.tsx
@@ -20,11 +20,11 @@ import AccessTimeOutlinedIcon from '@mui/icons-material/AccessTimeOutlined';
 import FormControl from "@mui/material/FormControl";
 import Highlighter from "react-highlight-words";
 import { useState } from "react";
-import { useSpanSearchStore } from "../../stores/spanSearchStore";
 
 import { useAvailableTags } from "../../api/availableTags";
 import { AvailableTag, TagGroup } from "../../types/availableTags";
 import { styles } from "./styles";
+import { recentlyUsedKeysStore } from "../../stores/recentlyUsedKeysStore";
 
 export type TagSelectorProps = {
   value: AvailableTag | null;
@@ -45,7 +45,7 @@ export const TagSelector = ({
     (page) => page.Tags
   );
 
-  const recentlyUsedKeys = useSpanSearchStore((state) => state.recentlyUsedKeysState.recentlyUsedKeys);
+  const recentlyUsedKeys = recentlyUsedKeysStore((state) => state.recentlyUsedKeysState.recentlyUsedKeys);
 
   let allTagsOptions: AvailableTag[] | undefined = undefined;
   if (availableTagsOptions) {
